Implement Deref for the Tracer
This makes the `Tracer` wrapper for a phy device a drop-in replacement
for the encapsulated device. All reference based methods of the traced
device stay callable, only those that take self by value need an
explicit unwrap with `into_inner`. There are no expected collisions with
the methods of the inner device as `into_inner` is the only method
provided by tracer.