Fix memory unsafety in socket ref
The previous version contained an aliasing mutable borrow in Drop and
the return of into_inner, that was only allowed as a backwards
compatible decision from a previous bug in the Rust borrow checker. As a
side effect of this change, the size of Ref_ decreased to the size of a
pointer due to an optimization on the layout of Option containing a
reference.